Create GetProductKeys.vbs
This script, GetProductKeys.VBS, is designed of course to grab the Windows and Office product keys from a local system. It supports Office versions 2002 (XP) - 2013, multiple versions of Office installed on the same system, and is 32bit and 64bit compatible. After keys are retrieved of course you have the option to save them to a text file. Specifically tested on Windows XP with Office 2007 and Windows 8 with Office 2010 and verified accuracy with Nirsoft's Produkey.